requirements.txt
torch==2.1.0
transformers==4.35.0
pdf2image==1.16.3
Pillow==10.0.0
pytesseract==0.3.10
jsonschema>=4.17.3
